+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 
 ### Added
 - Added use case to vtkh data adaptor for a mesh with a uniform cell set and explicit coordinates.
+- Added a compressed color table format.
+- Added action options relating to logging functionality including `open_log`, `flush_log`, and `close_log` to toggle logging as well as `set_log_threshold` and `set_echo_threshold` to control logging and standard output levels.
 - Added a new unified logging infrastructure.
 - Added support for unstructured topologies with mixed elements types (for example, hexs and tets).
 - Added support for `pyramid` and `wedge` elements.
@@ -25,6 +27,9 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 - Added `near_plane` and `far_plane` to the camera details provided in Ascent::info()
 - Added `add_mpi_ranks` and `add_domain_ids` filters for adding rank and domain fields to a mesh
 - Added `transform` filter, which allows you to rotate, scale, translate, mesh coordinates
+- Added python script in src/utilities/visit_session_converters to convert VisIt color table to Ascent actions color table
+- Added `fields` option to the project 2d to support scalar rendering of specific fields.
+- Added `dataset_bounds` option to the project 2d, which can be used instead of a full 3D camera specification
 
 ### Changed
 - Changed the replay utility's binary names such that `replay_ser` is now `ascent_replay` and `raplay_mpi` is now `ascent_replay_mpi`. This will help prevent potential name collisions with other tools that also have replay utilities. 
@@ -34,6 +39,7 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 - Resolved a bug where a sharing a coordset between multiple polytopal topologies would corrupt mesh processing.
 - Fixed a bug with Cinema resource output that could lead to corrupted html results.
 - Fixed a bug where controls for world and screen annotations where ignored in Cinema renders.
+- Fixed a bug in Uniform Grid Sampling and changed how ties for valid points are broken.
 
 ## [0.9.3] - Released 2024-05-11
 ### Preferred dependency versions for ascent@0.9.3